+++ frontend/src/Cabinet/components/Login/LoginTemplate.tsx
@@ -1,12 +1,8 @@
-import { Link } from "react-router-dom";
 import styled from "styled-components";
+import LoginCard from "@/Cabinet/components/Login/LoginCard";
 import { ReactComponent as LoginImg } from "@/Cabinet/assets/images/loginImg.svg";
-import { ReactComponent as LogoImg } from "@/Cabinet/assets/images/logo.svg";
-import LoginButtonGroup from "./LoginButtonGroup";
 
-const LoginTemplate = (props: { pageTitle: string; pageSubTitle: string }) => {
-  const { pageTitle, pageSubTitle } = props;
-
+const LoginTemplate = () => {
   return (
     <LoginPageStyled id="loginPage">
       <LeftSectionStyled className="leftLoginPage">
@@ -29,17 +25,7 @@ const LoginTemplate = (props: { pageTitle: string; pageSubTitle: string }) => {
         </BottomContentsStyled>
       </LeftSectionStyled>
       <RightSectionStyled className="rightLoginPage">
-        <LoginCardStyled className="modal">
-          <CardLogoStyled>
-            <LogoImg />
-          </CardLogoStyled>
-          <CardTitleBoxStyled>
-            <CardTitleStyled>{pageTitle}</CardTitleStyled>
-            <CardSubTitleStyled>{pageSubTitle}</CardSubTitleStyled>
-          </CardTitleBoxStyled>
-          <LoginButtonGroup />
-          <AGUURLSectionStyled to="/agu">A.G.U 유저라면?</AGUURLSectionStyled>
-        </LoginCardStyled>
+        <LoginCard />
       </RightSectionStyled>
     </LoginPageStyled>
   );
@@ -101,51 +87,6 @@ const RightSectionStyled = styled.section`
   justify-content: center;
   align-items: center;
   background-color: var(--sys-main-color);
-`;
-
-const LoginCardStyled = styled.div`
-  width: 350px;
-  height: 500px;
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  padding-top: 85px;
-  background-color: var(--bg-color);
-`;
-
-const CardLogoStyled = styled.div`
-  width: 70px;
-  height: 70px;
-  svg {
-    .logo_svg__currentPath {
-      fill: var(--sys-main-color);
-    }
-  }
-`;
-
-const CardTitleBoxStyled = styled.div`
-  text-align: center;
-  margin-top: 25px;
-`;
-
-const CardTitleStyled = styled.h1`
-  font-size: 2.5rem;
-  margin-bottom: 10px;
-  line-height: 2.5rem;
-`;
-
-const CardSubTitleStyled = styled.p`
-  color: var(--sys-main-color);
-`;
-
-const AGUURLSectionStyled = styled(Link)`
-  background-color: var(--bg-color);
-  color: var(--gray-line-btn-color);
-  margin-top: 10px;
-  margin-top: 24px;
-  text-decoration: underline;
-  font-size: 0.875rem;
-  font-weight: lighter;
 `;
 
 export default LoginTemplate;
