+++ frontend/src/Cabinet/pages/LoginPage.tsx
@@ -2,9 +2,7 @@ import "@/Cabinet/assets/css/loginPage.css";
 import LoginTemplate from "@/Cabinet/components/Login/LoginTemplate";
 
 const LoginPage = () => {
-  return (
-    <LoginTemplate pageTitle="Cabi" pageSubTitle="여러분의 일상을 가볍게" />
-  );
+  return <LoginTemplate />;
 };
 
 export default LoginPage;
